apply test file changes from review (2)
